more frontend/backend testing, removed outdated src directory
