+++ shoe-yelp/client/package.json
@@ -5,6 +5,7 @@
   "homepage": "https://flin1.github.io/shoe-yelp",
   "predeploy": "npm run build",
   "deploy": "gh-pages -d build",
+  "proxy": "http://localhost:3001",
   "dependencies": {
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
