+++ shoe-yelp/client/src/components/Navbar/index.css
@@ -1,11 +1,49 @@
+@import url(https://fonts.googleapis.com/css?family=Lato:400,300,100,700,900);
+
+h1{
+  font-family: 'Lato';
+  font-size: 2.8em;
+  padding: 10px 0;
+  font-weight: 800;
+}
+
+h3 {
+    font-family: 'Lato';
+}
+
+p {
+  font-size: 1.1em;
+  font-weight: 100;
+  font-family: 'Courier New', Courier, monospace;
+}
+
+.header {
+  width: 100%;
+  padding:60px 0 20px;
+  text-align: center;
+  background: #616161;
+  color: white;
+}
+#icon {
+    float:left;
+}
 #logo {
     height: 50px;
     width: 170px;
     margin: 20px;
 }
-#header {
-    display: inline-block;
+#search{
+    float: right;
+    margin: 20px;
 }
-#profile {
-    float:right;
+.profile {
+    padding-top: 10px;
+    float: right;
+    margin: 20px;
+}
+.login {
+  display: inline-block;
+}
+.signup {
+  display: inline-block;
 }
