+++ shoe-yelp/client/src/pages/about/about.css
@@ -1,3 +1,6 @@
 .body {
   text-align: center;
 }
+.about {
+  margin: 70px;
+}
