+++ shoe-yelp/client/src/pages/forum/forum.css
@@ -1,3 +1,12 @@
 .body {
   text-align: center;
 }
+.forum{
+  margin: 50px;
+  text-align: center;
+}
+.card {
+  float: auto;
+  padding: 20px;
+  display: inline-block;
+}
