+++ shoe-yelp/client/src/pages/home/home.css
@@ -1,0 +1,12 @@
+.body {
+    align-items: center;
+}
+.home {
+    margin: 70px;
+}
+#banner{
+    height: auto;
+    width: 100%;
+    background-size:contain;
+}
+
