--- shoe-yelp/client/src/pages/signup/signUpPage.css
@@ -1,96 +1,0 @@
-.pages{
-    list-style: none;
-    display: inline-block;
-    padding-right: 100px;
-  }
-  .search {
-    text-align: right;
-    padding-right: 100px;
-  }
-  .menu {
-    padding-top: 20px;
-    text-align: center;
-  }
-  .logo {
-    padding-top: 40px;
-    margin-left: 40px;
-    text-align: left;  
-  }
-  #logo_img {
-    height: 50px;
-    width:180px;
-  }
-  #title {
-    font-family: 'Antic Slab', serif;
-    text-transform: lowercase;
-  }
-  .link:hover {
-    text-decoration: none;
-  }
-  #home:hover {
-    text-decoration: none;
-  }
-  .intro {
-    text-align: left;
-    padding: 15%;
-  }
-  .events {
-    text-align: center;
-    padding: 20px;
-  }
-  .newsletter {
-    text-align: center;
-    padding: 20px;
-  }
-  .membership {
-    text-align: center;
-  }
-  .follow {
-    text-align: center;
-  }
-  .contain {
-    font-family: 'Abel', sans-serif;
-    padding: 20px;
-  }
-  .card {
-    /* Add shadows to create the "card" effect */
-    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
-    transition: 0.3s;
-    width: 20%;
-    display: inline-block;
-    padding: 20px;
-    margin-left: 20px;
-    margin-right: 20px;
-  }
-  
-  /* On mouse-over, add a deeper shadow */
-  .card:hover {
-    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
-  }
-  
-  /* Add some padding inside the card container */
-  .container {
-    padding-top: 10px;
-  }
-  h2 {
-    font-family: 'Antic Slab', serif;
-  }
-  
-  .label {
-    font-family: 'Abel';
-    font-size: 16px;
-  }
-  
-.mainContainer{
-  justify-content: center;
-  align-items: center;
-}
-.signUpButton {
-    margin-top: 10px;
-    margin-bottom: 10px;
-    border-width: 3px;
-    background-color: white;
-    border-color: black;
-    font-weight: 400px;
-    font-size: 20px;
-}
