+++ shoe-yelp/client/src/pages/trending/trending.css
@@ -1,3 +1,12 @@
 .body {
     text-align: center;
-  }
+}
+.trending {
+  margin: 50px;
+  text-align: center;
+}
+.card {
+  float: auto;
+  padding: 20px;
+  display: inline-block;
+}
