--- src/components/Navbar/NavbarElements.js
@@ -1,40 +1,0 @@
-import { NavLink as Link } from "react-router-dom";
-import styled from "styled-components";
-
-// pulled from GeeksforGeeks https://www.geeksforgeeks.org/how-to-create-a-multi-page-website-using-react-js/
-export const Nav = styled.nav`
-  height: 85px;
-  display: flex;
-  justify-content: space-between;
-  padding: 0.2rem calc((100vw - 1000px) / 2);
-  z-index: 12;
-`;
-  
-export const NavLink = styled(Link)`
-  color: #808080;
-  display: flex;
-  align-items: center;
-  text-decoration: none;
-  padding: 0 1rem;
-  height: 100%;
-  cursor: pointer;
-  &.active {
-    color: #000000;
-    text-decoration: none;
-  }
-`;
-  
-export const NavMenu = styled.div`
-  display: flex;
-  align-items: center;
-  margin-right: -24px;
-  /* Second Nav */
-  /* margin-right: 24px; */
-  /* Third Nav */
-  /* width: 100vw;
-white-space: nowrap; */
-  @media screen and (max-width: 768px) {
-    display: none;
-  }
-`;
-
