--- src/index.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-  
-ReactDOM.render(
-<React.StrictMode>
-    <App />
-</React.StrictMode>,
-document.getElementById('root')
-);
