--- src/pages/about/about.css
@@ -1,6 +1,0 @@
-.body {
-  text-align: center;
-}
-.about {
-  margin: 70px;
-}
