--- src/pages/contact/contact.css
@@ -1,15 +1,0 @@
-.body {
-  text-align: center;
-}
-#subheader {
-  margin: 50px;
-}
-.contactForm {
-  margin: 50px;
-}
-#name {
-  margin: 10px;
-}
-#email {
-  margin: 10px;
-}
