--- src/pages/forum/forum.css
@@ -1,12 +1,0 @@
-.body {
-  text-align: center;
-}
-.forum{
-  margin: 50px;
-  text-align: center;
-}
-.card {
-  float: auto;
-  padding: 20px;
-  display: inline-block;
-}
