--- src/pages/trending/trending.css
@@ -1,12 +1,0 @@
-.body {
-    text-align: center;
-}
-.trending {
-  margin: 50px;
-  text-align: center;
-}
-.card {
-  float: auto;
-  padding: 20px;
-  display: inline-block;
-}
